_6mans/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
discord.py==1.1.1
idna==2.8
multidict==4.5.2
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
